Fix aliasing issues
